refactor: refactor codebase to be more rust and diesel idiomatic
Separation of concern by separating the models in models folder and db
related stuff in the db folder.
